Refactor session no.2
Added a progress bar to the console output
Clean up the console output
Moved the helper functions to a new helper.h header file for inclusion
