+++ Input/ang_mmn.in
@@ -1,7 +1,7 @@
-2
-0
+1
+1
 200
-0.005
+0.00005
 0	0	0
-1.5	2.1	0.0	0.01	0.0	0.0	1	0.5	0	0.5	10	1000000000	0.01
-4.5	3.0	0.0	-0.01	0.0	0.0	1	0.5	0	0.5	10	1000000000	0.01
+1.5	2.1	0.00000000000001	0.01	0.0	0.0	1	0.5	1	0.5	10	1000000000	0.01
+4.5	3.0	0.00000000000001	-0.01	0.0	0.0	1	0.5	1	0.5	10	1000000000	0.01
