+++ Input/grv_test.in
@@ -1,9 +1,9 @@
 11
 0
-5
-1e-006
+20
+1e-004
 0	-9.81	0
-0.48	0.201	0.0	0.0	8.74357e-012	0.0	1	0.006	0	0.5	10	0.0003	0.01
+0.48	0.301	0.0	0.0	8.74357e-012	0.0	1	0.006	0	0.5	10	0.0003	0.01
 0.45	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
 0.46	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
 0.47	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
