+++ Input/grv_test2.in
@@ -1,6 +1,6 @@
 11
 0
-5
+1
 1e-006
 0	-9.81	0
 0.48	2.0001e-001	0.0	0.0	8.74357e-012	0.0	1	0.006	0	0.5	10	0.0003	0.01
