+++ Input/hopper.in
@@ -1,19 +1,19 @@
 511
-0
 1
+5
 1e-005
 0	-9.81	0
-0.45	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.46	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.47	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.48	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.49	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.5	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.51	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.52	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.53	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.54	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
-0.55	0.19	0	0	0	0	1	0.005	1	0.5	10	0.0003	0.01
+0.45	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.46	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.47	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.48	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.49	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.5	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.51	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.52	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.53	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.54	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
+0.55	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
 0.5525	0.2	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
 0.4475	0.2	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
 0.5575	0.21	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
