+++ Input/silo.in
@@ -1,6 +1,6 @@
 321
+0
 1
-100
 1e-005
 0	-9.81	0
 0.25	0.19	0	0	0	0	1	0.005	1	0.5	10	100000	0.01
