+++ Input/total_eng.in
@@ -1,6 +1,6 @@
 2
 0
-200
+1
 0.00005
 0	0	0
 1.5	3.0	0.0	0.01	0.0	0.0	1	0.5	0	0.5	10	1000000000	0.01
