+++ Input/window_box.in
@@ -1,4 +1,4 @@
-181
+280
 0
 1
 1e-6
